requirements.txt
treelib
rich
python-dotenv
tqdm
ijson